+++ Magic/CHANGELOG.md
@@ -24,6 +24,7 @@
  - Add `ModelEngine` action for detailed model control
  - Add `commandwand` and `commandspell` to set up command-bound items
  - The `transformable` mob property now applies to falling_block entities, and will prevent them from turning into blocks
+ - Add `Respawn` action, for respawning mobs that have died
 
 ## Changes